+++ CHANGELOG.md
@@ -1,3 +1,25 @@
+## [3.7.1](https://github.com/Sanofi-IADC/konviw/compare/v3.7.0...v3.7.1) (2023-05-30)
+
+
+### Bug Fixes
+
+* WEB-1927 Created observable timeout for external resources ([cf28615](https://github.com/Sanofi-IADC/konviw/commit/cf28615ee2eeb7759d41980bf3ec66fbdba794f3))
+
+# [3.7.0](https://github.com/Sanofi-IADC/konviw/compare/v3.6.3...v3.7.0) (2023-05-26)
+
+
+### Features
+
+* WEB-1891 Added conditional for header emoji ([9569acd](https://github.com/Sanofi-IADC/konviw/commit/9569acd1c621ccd00bb52061be47d6f2449b82c6))
+* WEB-1891 Migrated attachments into Confluence api v2 ([c7b44f1](https://github.com/Sanofi-IADC/konviw/commit/c7b44f1b188b71cd91ca551cc459aece8700a600))
+
+## [3.6.3](https://github.com/Sanofi-IADC/konviw/compare/v3.6.2...v3.6.3) (2023-05-24)
+
+
+### Bug Fixes
+
+* WEB-1915 Added implementation for emojis in headings ([c6c12ea](https://github.com/Sanofi-IADC/konviw/commit/c6c12eaaf9bed6ee7d5c2be6cc85c52f51a82959))
+
 ## [3.6.2](https://github.com/Sanofi-IADC/konviw/compare/v3.6.1...v3.6.2) (2023-05-15)